Readme
Description of the application and the motivation behind it The application will be in cooperation with Billigtlager.dk, which is a storage firm that currently is located in Næstved on Zealand. Billigtlager is aiming to open another storage warehouse in Vordingborg during the year 2023. Therefore, they have asked us to create an interactive application that can update contemporary with the already operating webpage to satisfy the increasing customer demand.

The application must contain the same specifications and features as the webpage. However, it would be encouraged to add more features such as a button that can connect with the automatic door system that is installed in the warehouse for the customers to enter the warehouse and their storage room.

List of prioritized requirements

As mentioned above the application must have the same features as the webpage which is as follows: • Frontpage

Providing standard information of the storage room and warehouse
Offers
How does it work (how to book and activate storage room) • A list containing:
Frontpage
Price information
Booking of storage room
About us
Division
Contact
Shopping bag
‘My Billigtlager’ (Customer page) • Price information
Listing the different types of rooms including prices
Listing all available storage rooms • About us
Storytelling • Contact
Providing information on various contact options
Make a speed dial button?
Able to send an email from the app? • Shopping bag
Will push the customer to the shopping bag with all the products • Customer page (only for active customers)
Provide personal information
Current payment method
Products
Etc.? • Booking of storage room
A list of all available storage rooms (clickable) -> pushes the customer to a check out page etc.
The above-listed requirements are all features that the web page currently provides to the customer. However, the possibility of executing orders and creating the payment method from the application will be the lowest-rated requirement. Extra features: • Changing division

(næstved vordingborg) • Button to open one of many doors (must send an SMS to do so)
User stories

As a user I want to be able to open the doors from the app
As a user I want to be able to see available storage rooms
As a user I want to be able to watch my current products
As a user I want to be able to log in to my profile
As a user I want to be able to see opening hours
As a user I want to be able to contact information
As a user I want to be able to see what sizes the storage rooms come in
As a user I want to be able to send emails through the app
As a user I want to be able to see address information
As a user I want to be able to read about the company storytelling
As a user I want to be able to set up and change the payment method
As a user I want to be able to deactivate membership
As a user I want to be able to shift between the divisions